refactor: factor out helpers
